refactor(dkg): align DKGSessionInfo structure with genesis contract (#13)
* refactor(dkg): align DKGSessionInfo structure with genesis contract

- Restructure DKGSessionInfo to contain nested DKGSessionMetadata
- Change ConfigV2Data threshold types from uint64 to uint128 for FixedPoint64 compatibility
- Add getDKGState() view function returning 4 fields for debugging
- Update DKGStartEvent to emit DKGSessionMetadata instead of full session info
- Update spec documentation to reflect all structural changes

The DKGSessionInfo now contains:
- metadata: DKGSessionMetadata (dealerEpoch, randomnessConfig, validator sets)
- startTimeUs: uint64
- transcript: bytes

+++ genesis-tool/src/genesis.rs
@@ -119,13 +119,13 @@ pub struct RandomnessConfigData {
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct ConfigV2Data {
     #[serde(rename = "secrecyThreshold")]
-    pub secrecy_threshold: u64,
+    pub secrecy_threshold: u128,
     
     #[serde(rename = "reconstructionThreshold")]
-    pub reconstruction_threshold: u64,
+    pub reconstruction_threshold: u128,
     
     #[serde(rename = "fastPathSecrecyThreshold")]
-    pub fast_path_secrecy_threshold: u64,
+    pub fast_path_secrecy_threshold: u128,
 }
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
@@ -205,9 +205,9 @@ sol! {
     }
     
     struct SolConfigV2Data {
-        uint64 secrecyThreshold;
-        uint64 reconstructionThreshold;
-        uint64 fastPathSecrecyThreshold;
+        uint128 secrecyThreshold;
+        uint128 reconstructionThreshold;
+        uint128 fastPathSecrecyThreshold;
     }
     
     struct SolRandomnessConfigData {
